Added custom hash function only for mongoose
