+++ lib/app.js
@@ -10,11 +10,10 @@ var mediator = require('fh-wfm-mediator/lib/mediator');
 var bodyParser = require('body-parser');
 var Promise = require('bluebird');
 var adminRouter = require('./app/admin/router')(mediator);
-var globalCollisionHandler = require('./data_collision_handlers/globalHandler');
+var initSync = require('./initSync');
 
 require('./env-var-check');
 require('fh-wfm-appform/lib/server')(mbaasApi);
-var wfmSync = require('fh-wfm-sync/lib/server');
 
 var authServiceGuid = process.env.WFM_AUTH_GUID;
 var securableEndpoints = [];
@@ -63,27 +62,6 @@ app.post('/cloud/:datasetId', function(req, res) {
 var excludedPaths = ['/authpolicy', '/file'];
 app.use(require('fh-wfm-user/lib/cloud/middleware/validateSession')(mediator, mbaasApi,excludedPaths));
 
-var syncOptions =   config.get("syncOptions");
-syncOptions.dataCollisionHandler = globalCollisionHandler;
-
-/**
- *
- * We are using mongoose models that have a `updatedAt` field.
- *
- * This can be used to detect if the record has changed.
- *
- * @param record
- */
-syncOptions.hashFunction = function generateHash(record) {
-  return record.updatedAt ? new Date(record.updatedAt).getTime() : undefined;
-};
-
-//Initialising the fh-wfm-sync data sets for each module
-wfmSync.init(mediator, mbaasApi, 'workorders', syncOptions);
-wfmSync.init(mediator, mbaasApi, 'result', syncOptions);
-wfmSync.init(mediator, mbaasApi, 'workflows', syncOptions);
-wfmSync.init(mediator, mbaasApi, 'messages', syncOptions);
-
 require('fh-wfm-message/lib/server')(mediator, app, mbaasApi);
 require('fh-wfm-result/lib/cloud')(mediator);
 require('fh-wfm-user/lib/cloud')(mediator, app, authServiceGuid);
@@ -97,6 +75,10 @@ var store = require("./storage/storage-init");
 module.exports = function() {
   //make sure that all app modules finish setting up
   var storagePromise = store.connect().then(function() {
+
+    //Initialising the sync handlers for this application
+    initSync(mediator, mbaasApi, store.usingMongooseStore());
+
     console.log("Initializing storage for objects");
     return Promise.all([
       require('./app/message')(mediator),
