+++ lib/storage/storage-init.js
@@ -93,5 +93,8 @@ function disconnect() {
 module.exports = {
   init: initStore,
   connect: connect,
-  disconnect: disconnect
+  disconnect: disconnect,
+  usingMongooseStore: function() {
+    return useMongooseStore === true;
+  }
 };
